put functionality to the separate app
