+++ images/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from profiles.models import Profile
+
+
+class BannerSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Profile
+        fields = ('banner_image',)
