+++ images/views.py
@@ -1,0 +1,16 @@
+from rest_framework.generics import (
+    RetrieveUpdateAPIView,
+)
+from rest_framework.parsers import MultiPartParser, FormParser
+from rest_framework.permissions import (
+    IsAuthenticatedOrReadOnly,
+)
+
+from profiles.models import Profile
+from .serializers import BannerSerializer
+
+class BannerChangeAPIView(RetrieveUpdateAPIView):
+    permission_classes = (IsAuthenticatedOrReadOnly,)
+    serializer_class = BannerSerializer
+    parser_classes = (MultiPartParser, FormParser)
+    queryset = Profile.objects.all()
