+++ profiles/urls.py
@@ -9,7 +9,6 @@ from .views import (
     CategoryDetail,
     ActivityDetail,
     ProfileViewCreate,
-    BannerChangeAPIView,
 )
 
 
@@ -38,5 +37,4 @@ urlpatterns = [
     path("activities/", ActivityList.as_view(), name="activity_list"),
     path("categories/<pk>/", CategoryDetail.as_view(), name="category-detail"),
     path("activities/<pk>/", ActivityDetail.as_view(), name="activity-detail"),
-    path("banner/<pk>/", BannerChangeAPIView.as_view(), name="banner_change")
 ]
